+++ src/components/maps/index.ts
@@ -1,3 +1,4 @@
+export { default as MapContainer } from './MapContainer';
 export { default as Map } from './Map';
 export { default as Marker } from './Marker';
 export { default as InfoBox } from './InfoBox';
