+++ src/types/map.ts
@@ -13,3 +13,10 @@ export type Info = {
   line: number;
   setLocation: React.Dispatch<React.SetStateAction<LocationType>>;
 };
+
+export type GoogleMapProps = {
+  children: React.ReactNode;
+  location: LocationType;
+  map: google.maps.Map | null;
+  setMap: React.Dispatch<React.SetStateAction<google.maps.Map | null>>;
+};
